🐄🎊 Checkpoint ./komutlar/deneme.js:3756271/8085 ./komutlar/instagram.js:3756271/1754

+++ komutlar/instagram.js
@@ -1,27 +1,34 @@
-const Discord = module.require('discord.js');
-const ayarlar = require('../ayarlar.json')
-const prefix = ayarlar.prefix
-const client = new Discord.Client();
-module.exports.run = async (client, message, args) => {
+const Discord = require('discord.js');
+exports.run = (client, message, args) => {
+    message.delete()
+  if (!message.guild) {
+  const ozelmesajuyari = new Discord.RichEmbed()
+  .setColor(0xFF0000)
+  .setTimestamp()
+  .setAuthor(message.author.username, message.author.avatarURL)
+  .addField('⚠ Uyarı ⚠', 'Bu  komutu özel mesajlarda kullanamazsın.');
+  return message.author.sendEmbed(ozelmesajuyari); }
+  let guild = message.guild;
+  let reason = args.slice(0).join(' ');
+  if (reason.length < 1) return message.reply('İnstagram Adresini Yazmayı Unuttunuz.');
   message.delete();
-  let instagram = args.slice(1).join(' ');
-  const inst = new Discord.RichEmbed()
+  message.reply('Mesajını Gönderdim.')
+  const embed = new Discord.RichEmbed()
   .setColor('RANDOM')
-  .setDescription(instagram);
+  .setDescription(reason);
+      let insta = message.guild.channels.find(`name`, "📸ɪɴsᴛᴀɢʀᴀᴍ");    
+        message.guild.channels.get(insta.id).send(embed)
+};
 
-
-    let insta = message.guild.channels.find(`name`, "📸ɪɴsᴛᴀɢʀᴀᴍ");    
-        message.guild.channels.get(insta.id).send(inst)
-  
-}
 exports.conf = {
   enabled: true,
-  guildOnly: true,
-  aliases: ['instagram'],
-  permLevel: 0
+  guildOnly: false,
+  aliases: ['insta'],
+  permlevel: 4
 };
+
 exports.help = {
-  name: 'instagram',
-  description: '!!instagram [KULLANICI ADI].',
-  usage: 'instagram'
+  name: 'insta',
+  description: 'Bir kullanÄ±cÄ±ya Ã¶zel mesaj yollar.',
+  usage: 'insta'
 };
